Refactor to move public to private types
